requirements.txt
praw
flask
selenium
python-dotenv
Pillow
pytesseract
gtts